microweb/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.6
requests==2.19.1
python-consul==1.1.0
consulservicefinder==0.0.2